Sell - Refactoring + flow
