+++ src/exchange/hw-app-exchange/Exchange.js
@@ -4,25 +4,33 @@ import { BigNumber } from "bignumber.js";
 import { TransportStatusError } from "@ledgerhq/errors";
 import invariant from "invariant";
 
+const TRANSACTION_TYPES = {
+  SWAP: 0x00,
+  SELL: 0x01,
+};
+type TransactionType = $Values<typeof TRANSACTION_TYPES>;
+
 const START_NEW_TRANSACTION_COMMAND: number = 0x03;
 const SET_PARTNER_KEY_COMMAND: number = 0x04;
 const CHECK_PARTNER_COMMAND: number = 0x05;
 const PROCESS_TRANSACTION_RESPONSE: number = 0x06;
 const CHECK_TRANSACTION_SIGNATURE: number = 0x07;
 const CHECK_PAYOUT_ADDRESS: number = 0x08;
+const CHECK_ASSET_IN: number = 0x08;
 const CHECK_REFUND_ADDRESS: number = 0x09;
 const SIGN_COIN_TRANSACTION: number = 0x0a;
 
 const maybeThrowProtocolError = (result: Buffer): void => {
-  invariant(result.length >= 2, "SwapTransport: Unexpected result length");
+  invariant(result.length >= 2, "ExchangeTransport: Unexpected result length");
   const resultCode = result.readUInt16BE(result.length - 2);
   if (resultCode !== 0x9000) {
     throw new TransportStatusError(resultCode);
   }
 };
 
-export default class Swap {
+export default class Exchange {
   transport: Transport<*>;
+  transactionType: TransactionType;
   allowedStatuses: Array<number> = [
     0x9000,
     0x6a80,
@@ -36,7 +44,8 @@ export default class Swap {
     0x9d1a,
   ];
 
-  constructor(transport: Transport<*>) {
+  constructor(transport: Transport<*>, transactionType: TransactionType) {
+    this.transactionType = transactionType;
     this.transport = transport;
   }
 
@@ -45,11 +54,15 @@ export default class Swap {
       0xe0,
       START_NEW_TRANSACTION_COMMAND,
       0x00,
-      0x00,
+      this.transactionType,
       Buffer.alloc(0),
       this.allowedStatuses
     );
     maybeThrowProtocolError(result);
+
+    if (this.transactionType === TRANSACTION_TYPES.SELL) {
+      return result.subarray(0, 32).toString("base64");
+    }
 
     return result.toString("ascii", 0, 10);
   }
@@ -59,7 +72,7 @@ export default class Swap {
       0xe0,
       SET_PARTNER_KEY_COMMAND,
       0x00,
-      0x00,
+      this.transactionType,
       partnerNameAndPublicKey,
       this.allowedStatuses
     );
@@ -72,7 +85,7 @@ export default class Swap {
       0xe0,
       CHECK_PARTNER_COMMAND,
       0x00,
-      0x00,
+      this.transactionType,
       signatureOfPartnerData,
       this.allowedStatuses
     );
@@ -96,7 +109,7 @@ export default class Swap {
       0xe0,
       PROCESS_TRANSACTION_RESPONSE,
       0x00,
-      0x00,
+      this.transactionType,
       bufferToSend,
       this.allowedStatuses
     );
@@ -109,7 +122,7 @@ export default class Swap {
       0xe0,
       CHECK_TRANSACTION_SIGNATURE,
       0x00,
-      0x00,
+      this.transactionType,
       transactionSignature,
       this.allowedStatuses
     );
@@ -139,9 +152,11 @@ export default class Swap {
 
     let result: Buffer = await this.transport.send(
       0xe0,
-      CHECK_PAYOUT_ADDRESS,
+      this.transactionType === TRANSACTION_TYPES.SWAP
+        ? CHECK_PAYOUT_ADDRESS
+        : CHECK_ASSET_IN,
       0x00,
-      0x00,
+      this.transactionType,
       bufferToSend,
       this.allowedStatuses
     );
@@ -185,7 +200,7 @@ export default class Swap {
       0xe0,
       SIGN_COIN_TRANSACTION,
       0x00,
-      0x00,
+      this.transactionType,
       Buffer.alloc(0),
       this.allowedStatuses
     );
