+++ src/exchange/sell/types.js
@@ -1,0 +1,13 @@
+// @flow
+
+import type { Transaction } from "../../types";
+
+export type InitSellResult = {
+  transaction: Transaction,
+};
+
+export type SellRequestEvent =
+  | { type: "init-sell-requested" }
+  | { type: "init-sell-get-transaction-id", value: string }
+  | { type: "init-sell-error", error: Error }
+  | { type: "init-sell-result", initSellResult: InitSellResult };
