+++ src/exchange/swap/getProviders.js
@@ -1,11 +1,11 @@
 // @flow
 
 import type { GetProviders } from "./types";
-import network from "../network";
+import network from "../../network";
 import { getSwapAPIBaseURL } from "./";
-import { getEnv } from "../env";
+import { getEnv } from "../../env";
 import { mockGetProviders } from "./mock";
-import { SwapNoAvailableProviders } from "../errors";
+import { SwapNoAvailableProviders } from "../../errors";
 
 const getProviders: GetProviders = async () => {
   if (getEnv("MOCK")) return mockGetProviders();
