--- src/sell/mock.js
@@ -1,45 +1,0 @@
-// @flow
-
-import type {
-  GetMultipleStatus,
-} from "./types";
-import type { Transaction } from "../types";
-import { Observable, of } from "rxjs";
-
-export const mockInitSell = (
-  transaction: Transaction
-): Observable<SwapRequestEvent> => {
-  return of({
-    type: "init-swap-result",
-    initSwapResult: {
-      transaction,
-      swapId: "mockedSwapId",
-    },
-  });
-};
-
-export const mockGetProviders: GetProviders = async () => {
-  //Fake delay to show loading UI
-  await new Promise((r) => setTimeout(r, 800));
-
-  return [
-    {
-      provider: "changelly",
-      supportedCurrencies: [
-        "bitcoin",
-        "litecoin",
-        "ethereum",
-        "tron",
-        "ethereum/erc20/omg",
-        "ethereum/erc20/0x_project",
-        "ethereum/erc20/augur",
-      ],
-    },
-  ];
-};
-
-export const mockGetStatus: GetMultipleStatus = async (statusList) => {
-  //Fake delay to show loading UI
-  await new Promise((r) => setTimeout(r, 800));
-  return statusList.map((s) => ({ ...s, status: "failed" }));
-};
